+++ tests/test_tasks.py
@@ -1016,41 +1016,23 @@ class DeliveryTasksTest(asynctest.TestCase):
         self.mocked.get(url, status=200, body=json.dumps(body))
 
     def _telemetry_mock_nightly_query(self, body=None):
+        record = {
+            "latest_query_data_id": 5678,
+            "id": 40197,
+            "name": "Uptake Firefox NIGHTLY"
+        }
         if body is None:
-            body = [{
-                "latest_query_data_id": 5678,
-                "id": 40197,
-                "name": "Uptake Firefox NIGHTLY 57.0a1 20170920"
-            }]
+            body = [record]
 
-        url = ("{}/api/queries/search?q=Uptake+Firefox+NIGHTLY+57.0a1+"
-               "%2820170920220431%2C+20170920111019%2C+20170920100426%29&include_drafts=true")
+        url = ("{}/api/queries/search?q=Uptake+Firefox+NIGHTLY&include_drafts=true")
         url = url.format(telemetry.TELEMETRY_SERVER)
         self.mocked.get(url, status=200, body=json.dumps(body))
 
+        url = "{}/api/queries/40197".format(telemetry.TELEMETRY_SERVER)
+        self.mocked.post(url, status=200, body=json.dumps(record))
+
     def _telemetry_mock_query_result(self, body):
         url = '{}/api/query_results/5678'.format(telemetry.TELEMETRY_SERVER)
-        self.mocked.get(url, status=200, body=json.dumps(body))
-
-    def _telemetry_mock_nightly_build_ids(self, body=None):
-        url = '{}/api/queries/{}'
-        url = url.format(telemetry.TELEMETRY_SERVER, telemetry.NIGHTLY_BUILD_IDS["57.0a1"])
-        self.mocked.get(url, status=200, body=json.dumps({
-            "latest_query_data_id": 1234
-        }))
-
-        if body is None:
-            body = {
-                "query_result": {"data": {"rows": [
-                    {"build_id": "20170920220431"},
-                    {"build_id": "20170920111019"},
-                    {"build_id": "20170920100426"},
-                    {"build_id": "20170919220202"},
-                    {"build_id": "20170919110626"}
-                ]}}
-            }
-
-        url = '{}/api/query_results/1234'.format(telemetry.TELEMETRY_SERVER)
         self.mocked.get(url, status=200, body=json.dumps(body))
 
     async def test_telemetry_update_uptake_tasks_returns_error_for_previous_nightly(self):
@@ -1058,22 +1040,8 @@ class DeliveryTasksTest(asynctest.TestCase):
         assert received["status"] == Status.MISSING.value
         assert received["message"] == "Telemetry update-parquet metrics landed in Firefox Quantum"
 
-    async def test_telemetry_update_uptake_tasks_returns_error_for_unsupported_nightly(self):
-        with pytest.raises(TaskError) as excinfo:
-            await telemetry.update_parquet_uptake('firefox', '57.0a2')
-        assert str(excinfo.value) == 'Please configure Build IDs query for 57.0a2'
-
-    async def test_telemetry_update_uptake_tasks_returns_error_for_unavailable_query(self):
-        url = '{}/api/queries/{}'
-        url = url.format(telemetry.TELEMETRY_SERVER, telemetry.NIGHTLY_BUILD_IDS["57.0a1"])
-        self.mocked.get(url, status=502)
-
-        with pytest.raises(TaskError) as excinfo:
-            await telemetry.update_parquet_uptake('firefox', '57.0a1')
-        assert str(excinfo.value) == 'Query 40223 unavailable (HTTP 502)'
-
     async def test_telemetry_update_uptake_tasks_returns_incomplete_for_no_result(self):
-        self._telemetry_mock_nightly_build_ids()
+        self._mock_buildhub_search()
         self._telemetry_mock_nightly_query([{
             "latest_query_data_id": None,
             "id": 40197,
@@ -1085,7 +1053,7 @@ class DeliveryTasksTest(asynctest.TestCase):
         assert received["message"] == ("Query still processing.")
 
     async def test_telemetry_update_uptake_tasks_returns_error_for_empty_results(self):
-        self._telemetry_mock_nightly_build_ids()
+        self._mock_buildhub_search()
         self._telemetry_mock_nightly_query()
         self._telemetry_mock_query_result({
             "query_result": {"data": {"rows": []}}
@@ -1095,21 +1063,8 @@ class DeliveryTasksTest(asynctest.TestCase):
         assert received["status"] == Status.ERROR.value
         assert received["message"] == ("No result found for your query.")
 
-    async def test_telemetry_update_uptake_tasks_returns_error_for_unavailable_query_results(self):
-        url = '{}/api/queries/{}'
-        url = url.format(telemetry.TELEMETRY_SERVER, telemetry.NIGHTLY_BUILD_IDS["57.0a1"])
-        self.mocked.get(url, status=200, body=json.dumps({
-            "latest_query_data_id": 1234
-        }))
-        url = '{}/api/query_results/1234'.format(telemetry.TELEMETRY_SERVER)
-        self.mocked.get(url, status=502)
-
-        with pytest.raises(TaskError) as excinfo:
-            await telemetry.update_parquet_uptake('firefox', '57.0a1')
-        assert str(excinfo.value) == 'Query Result 1234 unavailable (HTTP 502)'
-
     async def test_telemetry_update_uptake_tasks_returns_incomplete_for_low_nightly_uptake(self):
-        self._telemetry_mock_nightly_build_ids()
+        self._mock_buildhub_search()
         self._telemetry_mock_nightly_query()
         self._telemetry_mock_query_result({
             "query_result": {"data": {"rows": [
@@ -1120,11 +1075,10 @@ class DeliveryTasksTest(asynctest.TestCase):
         received = await telemetry.update_parquet_uptake('firefox', '57.0a1')
         assert received["status"] == Status.INCOMPLETE.value
         assert received["message"] == ("Telemetry uptake for version 57.0a1 "
-                                       "(20170920220431, 20170920111019, 20170920100426) "
-                                       "is 45.32%")
+                                       "(20171009192146) is 45.32%")
 
     async def test_telemetry_update_uptake_tasks_should_ignore_copied_queries(self):
-        self._telemetry_mock_nightly_build_ids()
+        self._mock_buildhub_search()
         self._telemetry_mock_nightly_query([{
             "latest_query_data_id": 123456789,
             "id": 40198,
@@ -1143,11 +1097,10 @@ class DeliveryTasksTest(asynctest.TestCase):
         received = await telemetry.update_parquet_uptake('firefox', '57.0a1')
         assert received["status"] == Status.EXISTS.value
         assert received["message"] == ("Telemetry uptake for version 57.0a1 "
-                                       "(20170920220431, 20170920111019, 20170920100426) "
-                                       "is 65.43%")
+                                       "(20171009192146) is 65.43%")
 
     async def test_telemetry_update_uptake_tasks_returns_exists_for_high_nightly_uptake(self):
-        self._telemetry_mock_nightly_build_ids()
+        self._mock_buildhub_search()
         self._telemetry_mock_nightly_query()
         url = '{}/api/query_results/5678'.format(telemetry.TELEMETRY_SERVER)
         self.mocked.get(url, status=200, body=json.dumps({
@@ -1159,11 +1112,10 @@ class DeliveryTasksTest(asynctest.TestCase):
         received = await telemetry.update_parquet_uptake('firefox', '57.0a1')
         assert received["status"] == Status.EXISTS.value
         assert received["message"] == ("Telemetry uptake for version 57.0a1 "
-                                       "(20170920220431, 20170920111019, 20170920100426) "
-                                       "is 65.43%")
+                                       "(20171009192146) is 65.43%")
 
     async def test_telemetry_update_uptake_tasks_returns_missing_for_no_search_query(self):
-        self._telemetry_mock_nightly_build_ids()
+        self._mock_buildhub_search()
         self._telemetry_mock_nightly_query()
         url = '{}/api/query_results/5678'.format(telemetry.TELEMETRY_SERVER)
         self.mocked.get(url, status=404)
@@ -1202,7 +1154,6 @@ class DeliveryTasksTest(asynctest.TestCase):
 
     async def test_telemetry_update_uptake_creates_the_query_if_not_found_for_nightly(self):
         self._mock_buildhub_search()
-        self._telemetry_mock_nightly_build_ids()
         self._telemetry_mock_nightly_query([])
 
         url = '{}/api/queries'.format(telemetry.TELEMETRY_SERVER)
@@ -1218,27 +1169,8 @@ class DeliveryTasksTest(asynctest.TestCase):
         received = await telemetry.update_parquet_uptake('firefox', '57.0a1')
         assert received["status"] == Status.INCOMPLETE.value
         assert received["message"] == (
-            "Telemetry uptake calculation for version 57.0a1 "
-            "(20170920220431, 20170920111019, 20170920100426) is in progress"
+            "Telemetry uptake calculation for version 57.0a1 (20171009192146) is in progress"
         )
-
-    async def test_telemetry_update_uptake_creates_the_query_if_null_body(self):
-        url = '{}/api/queries/{}'
-        url = url.format(telemetry.TELEMETRY_SERVER, telemetry.NIGHTLY_BUILD_IDS["57.0a1"])
-        self.mocked.get(url, status=200, body=json.dumps({}))
-
-        with pytest.raises(TaskError) as excinfo:
-            await telemetry.update_parquet_uptake('firefox', '57.0a1')
-        assert str(excinfo.value) == "Couldn't find any build matching."
-
-    async def test_telemetry_update_uptake_creates_the_query_if_no_results(self):
-        self._telemetry_mock_nightly_build_ids({
-            "query_result": {"data": {"rows": []}}
-        })
-
-        with pytest.raises(TaskError) as excinfo:
-            await telemetry.update_parquet_uptake('firefox', '57.0a1')
-        assert str(excinfo.value) == "Couldn't find any build matching."
 
     async def test_telemetry_update_uptake_creates_the_query_if_not_found_for_release(self):
         self._mock_buildhub_search()