+++ pollbot/tasks/bedrock.py
@@ -70,6 +70,12 @@ async def security_advisories(product, version):
             status = build_version_id(last_release) >= build_version_id(version)
             message = ("Security advisories for release were "
                        "published up to version {}".format(last_release))
+
+            version_title = "#firefox{}".format(version.split('.')[0])
+            if status and not d(version_title):
+                status = Status.INCOMPLETE
+                message += " but nothing was published for {}".format(version_title)
+
             return build_task_response(status, url, message)