+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ CHANGELOG
 
 - Reuse the same Nightly query for Telemetry Update Parquet. (#141)
 - Add bouncer checks and endpoints (#147)
+- Switch to Telemetry Athena Data Source (#155)
 
 
 0.5.0 (2017-11-06)