+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ CHANGELOG
 - Add bouncer checks and endpoints (#147)
 - Switch to Telemetry Athena Data Source (#155)
 - Add an indication about Crash-Stats 24h latency (#156)
+- Fix get_version_from_filename for all locales (#157)
 
 
 0.5.0 (2017-11-06)