+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ CHANGELOG
 - Always expect a major version security advisory title for release and ESR (#150)
 - Switch to Telemetry Athena Data Source (#155)
 - Add an indication about Crash-Stats 24h latency (#156)
+- Fix get_version_from_filename for all locales (#157)
 
 0.5.0 (2017-11-06)
 ------------------