+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ CHANGELOG
 
 - Reuse the same Nightly query for Telemetry Update Parquet (#141)
 - Add bouncer checks and endpoints (#147)
+- Handle case when Download links return a 504 instead of a 302 (#152)
 - Always expect a major version security advisory title for release and ESR (#150)
 - Switch to Telemetry Athena Data Source (#155)
 - Add an indication about Crash-Stats 24h latency (#156)